==> scripts/02 create_birthnode_awc_lanes.ipynb <==
This script helps identify the AWC closest to each birth node. We will use the following steps - 
1. Cross-join the birth-node and AWC locations df to get all unique birth-node-AWC combination (lane)
2. Calculate the distance for each lane
3. Eliminate all lanes longer than 20 km. The threshold is arbitrarily selected to reduce the size of the dataframe
4. Sort the lanes in ascending order (shortest lanes on top)
5. Drop duplicates at birth-node level, keeping only the first record, which gives us the closest AWC to the birth-node

import pandas as pd
from pathlib import Path
path_data = Path.cwd().parent / 'data' 

births_df = pd.read_csv(path_data / '01 out_births_cleaned_district.csv')
awc_locations_df = pd.read_csv(path_data / '02 in_awc_locations.csv')

phc_locations_df = pd.read_csv(path_data / '01 out_phc_locations_geocoded.csv')

# phc_locations_df = phc_locations_df[phc_locations_df['Primary Health Centre Name']=='Primary Health Centre'].reset_index(drop = True)

STEP 1

from math import radians, cos, sin, asin, sqrt

def haversine(lon1, lat1, lon2, lat2):
    """
    Calculate the great circle distance in kilometers between two points 
    on the earth (specified in decimal degrees)
    """
    # convert decimal degrees to radians 
    lon1, lat1, lon2, lat2 = map(radians, [lon1, lat1, lon2, lat2])

    # haversine formula 
    dlon = lon2 - lon1 
    dlat = lat2 - lat1 
    a = sin(dlat/2)**2 + cos(lat1) * cos(lat2) * sin(dlon/2)**2
    c = 2 * asin(sqrt(a)) 
    r = 6371 # Radius of earth in kilometers. Use 3956 for miles. Determines return value units.
    return c * r

# Data cleaning and prep dataframes for cross-join
births_df['Location ID'] = births_df.index

births_df['key'] = 1
awc_locations_df['key'] = 1

births_df = births_df.rename(columns = {'Latitude' : 'Birth Latitude', 'Longitude': 'Birth Longitude'})
awc_locations_df = awc_locations_df.rename(columns = {'Latitude' : 'AWC Latitude', 'Longitude' : 'AWC Longitude'})

birth_awc_lanes_df = births_df.merge(awc_locations_df, on='key', how='outer')

phc_locations_df['key'] = 1
phc_locations_df = phc_locations_df.rename(columns = {'Latitude' : 'PHC Latitude', 'Longitude' : 'PHC Longitude'})
birth_phc_lanes_df = births_df.merge(phc_locations_df, on='key', how='outer')

awc_phc_lanes_df = awc_locations_df.merge(phc_locations_df, on = 'key', how = 'outer')

STEP 2

birth_awc_lanes_df['Distance'] = birth_awc_lanes_df.apply(lambda x: haversine(lon1 = x['Birth Longitude'], lat1 = x['Birth Latitude'],
                                                         lon2 = x['AWC Longitude'], lat2 = x['AWC Latitude']), axis = 1)

birth_phc_lanes_df['Distance'] = birth_phc_lanes_df.apply(lambda x: haversine(lon1 = x['Birth Longitude'], lat1 = x['Birth Latitude'],
                                                         lon2 = x['PHC Longitude'], lat2 = x['PHC Latitude']), axis = 1)

awc_phc_lanes_df['Distance'] = awc_phc_lanes_df.apply(lambda x: haversine(lon1 = x['AWC Longitude'], lat1 = x['AWC Latitude'],
                                                         lon2 = x['PHC Longitude'], lat2 = x['PHC Latitude']), axis = 1)

STEP 3

birth_awc_lanes_df = birth_awc_lanes_df[birth_awc_lanes_df['Distance']<=20]

birth_phc_lanes_df = birth_phc_lanes_df[birth_phc_lanes_df['Distance']<=50]

awc_phc_lanes_df = awc_phc_lanes_df[awc_phc_lanes_df['Distance']<=100]

STEP 4

birth_awc_lanes_df = birth_awc_lanes_df.sort_values(by=['Location ID', 'Distance'], ascending = [True, True])
birth_awc_lane_assignment_df = birth_awc_lanes_df.drop_duplicates('Location ID', keep='first')

birth_phc_lanes_df = birth_phc_lanes_df.sort_values(by=['Location ID', 'Distance'], ascending = [True, True])
birth_phc_lane_assignment_df = birth_phc_lanes_df.drop_duplicates('Location ID', keep='first')

awc_phc_lanes_df = awc_phc_lanes_df.sort_values(by=['AWC_ID', 'Distance'], ascending = [True, True])
awc_phc_lane_assignment_df = awc_phc_lanes_df.drop_duplicates('AWC_ID', keep='first')

STEP 5

birth_awc_lane_assignment_df.to_csv(path_data / '02 out_birth_awc_lane_assignment.csv', index = False)

birth_phc_lane_assignment_df.to_csv(path_data / '02 out_birth_phc_lane_assignment.csv', index = False)

awc_phc_lane_assignment_df.to_csv(path_data / '02 out_awc_phc_lane_assignment.csv', index = False)
