==> scripts/06 Validate Mathematical Demand Equation.ipynb <==
import numpy as np

import pandas as pd

from pathlib import Path

import matplotlib.pyplot as plt 

path_data = Path.cwd().parent / 'data'

birth_awc_df = pd.read_csv(path_data / '02 out_birth_awc_lane_assignment.csv')

birth_phc_df = pd.read_csv(path_data / '02 out_birth_phc_lane_assignment.csv')

vaccine_details_df = pd.read_csv(path_data / '03 in_vaccine_details.csv')

vaccine_delay_geom_p = pd.read_csv(path_data / '03 out_vaccine_delay_geom_p.csv')

delta = 0.5 # Ratio of (distance from nearest AWC)/(distance from nearest PHC), where caregiver's choice probability is equal = 0.5

birth_awc_phc_df = pd.merge(birth_awc_df[['Location ID', 'AWC_ID', 'Distance', 'Births']].rename(columns = {'Distance':'AWC Distance'}), 
                            birth_phc_df[['Location ID', 'PHC_ID', 'Distance']].rename(columns = {'Distance': 'PHC Distance'}),
                            on = 'Location ID', how = 'left')

birth_awc_phc_df['AWC PHC Distance Ratio'] = birth_awc_phc_df['AWC Distance'] /  birth_awc_phc_df['PHC Distance']
birth_awc_phc_df['Normalized Distance Ratio'] = (birth_awc_phc_df['AWC PHC Distance Ratio'] - delta)/birth_awc_phc_df['AWC PHC Distance Ratio'].var()

birth_awc_phc_df['P_AWC'] = 1/(1 + np.exp(birth_awc_phc_df['Normalized Distance Ratio']))

vaccine_details_df['Wastage Adjustment'] = 1/(1 - np.array(vaccine_details_df['Wastage']))
vaccine_details_df['Wastage Adjusted Volume'] = np.ceil(vaccine_details_df['Average packed volume secondary package'] * vaccine_details_df['Wastage Adjustment'])

selected_awc = [1510219, 1510020, 1511016, 1511217, 1511061, 1510059, 1511012]

pois = lambda x: np.random.poisson(x/365, 730)

A = vaccine_details_df['Vaccine']
f = 28
vac_start = vaccine_details_df[['Vaccine', 'Start Week']]
alpha_i = 0.8 
beta_i = 0.1 
gamma_i = 0.4
p_vac = vaccine_delay_geom_p[['Vaccine', 'Vaccination Delay Geom p']]
delta_b = 0.2
vac_vol = vaccine_details_df[['Vaccine', 'Wastage Adjusted Volume']]

t_range = range(600, 730, 30)
# t_range = range(600, 601, 30)
awc_demand_volume_dict = {}

for awc in selected_awc:

    print(awc)
    
    birth_awc_iter_df = birth_awc_phc_df.loc[birth_awc_phc_df['AWC_ID'] == awc, :]
    
    births_array = np.array(list(map(pois, birth_awc_iter_df['Births'].values)))
    
    N = birth_awc_iter_df.shape[0]
    
    d = birth_awc_iter_df[['Location ID', 'AWC Distance', 'PHC Distance']].reset_index(drop=True)
    
    awc_choice_prob = birth_awc_iter_df[['Location ID', 'P_AWC']].reset_index(drop=True)

    awc_vaccine_volume = np.zeros((len(t_range), N))

    for n in range(N):
        print(n,end='\r')
        vac_seeking_probability = max((alpha_i - d.loc[n, ['AWC Distance','PHC Distance']].min() * beta_i),  gamma_i)
        for a in A:
            p_vac_value = p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p']
            vac_start_value = vac_start.loc[vac_start['Vaccine'] == a, 'Start Week']
            vac_vol_value = vac_vol.loc[vac_vol['Vaccine']==a,"Wastage Adjusted Volume"]
            for s in range(-4,1):
                for i in range(0, f):
                    t_start = 0
                    for t in t_range:
                        t_updated = t - vac_start_value * 7 + s*f - i
                        if births_array[n,t_updated]>0:
                            if s==0:
                                delay_cdf = 1 - (1 - p_vac_value)**(- s*f + i + 1)
                            else:
#                                 print(f's={s}')
#                                 print(f'i={i}')
#                                 delay_cdf = (1 - p_vac_value)**(-s * f) - (1 - p_vac_value)**(- s*f + i + 1)
#                                 print(f'Delay cdf 1: {delay_cdf}')

                                delay_cdf = (1 - p_vac_value)**(-(s+1) * f + i + 1) - (1 - p_vac_value)**(- s*f + i + 1)
#                                 print(f'Delay cdf 2: {delay_cdf}')
#                                 print(f'exp 1: {-(s-1) * f + i + 1}, exp 2: {(- s*f + i + 1)}')
                            awc_vaccine_volume[t_start, n] += (births_array[n,t_updated] * vac_seeking_probability * awc_choice_prob.loc[n, 'P_AWC'] * delay_cdf * (1-delta_b) + float(delta_b if s==0 else 0))*vac_vol_value
                        t_start += 1
                        
    awc_demand_volume_dict[awc] = awc_vaccine_volume

[(key, np.mean(np.sum(value, axis = 1))) for key,value in awc_demand_volume_dict.items()]

[(key, np.mean(np.sum(value, axis = 1))) for key,value in awc_demand_volume_dict.items()]

np.sum(awc_demand_volume_dict[1511016])

np.sum(awc_demand_volume_dict[1510020])

a = 'Rotavirus Vaccine 2'
n = 0
s = -1

delay_cdf = (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**(- s * f) - (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**( - s*f + i)

delay_cdf

x = float(delay_cdf * (1-delta_b))

y = delta_b if s==0 else 0

x+y

delay_cdf * (1-delta_b) + float(delta_b if s==0 else 0)

(births_array[n,599] * vac_seeking_probability * (delay_cdf * (1-delta_b) + delta_b if s==0 else 0))*vac_vol.loc[vac_vol['Vaccine']==a,"Wastage Adjusted Volume"]

awc_vaccine_volume = np.zeros((len(t_range), N))

vac_seeking_probability = (alpha_i - d.loc[n, 'Distance'] * beta_i).clip(min = gamma_i)


for i in range(1, f+1):  
    t_start = 0
    t_updated = t - vac_start.loc[vac_start['Vaccine'] == a, 'Start Week'] * 7 + s*f - i
    if i==1:
        print("Number of eligible children")
        print(births_array[n,(int(t_updated)-f):int(t_updated)])
        print(np.sum(births_array[n,(int(t_updated)-f):int(t_updated)]))
    if births_array[n,t_updated]>0:
        print(f"i = {i}")
        delay_cdf = (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**(- s * f) - (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**( - s*f + i)
        print(f"Delay cdf = {float(delay_cdf)}")
        awc_vaccine_volume[t_start, n] += (births_array[n,t_updated] * vac_seeking_probability * (delay_cdf * (1-delta_b) + float(delta_b if s==0 else 0)))*vac_vol.loc[vac_vol['Vaccine']==a,"Wastage Adjusted Volume"]
        print(f"Vaccine volume = {awc_vaccine_volume[t_start, n]}")
    t_start += 1

813

vac_vol.loc[vac_vol['Vaccine']==a,"Wastage Adjusted Volume"]

np.sum(births_array[:,400:500])

np.mean(np.sum(awc_vaccine_volume, axis =1))

# n=0
# a='BCG'
# s=0
# t=700
# i=1

# t_updated = t - vac_start.loc[vac_start['Vaccine'] == a, 'Start Week'] * 7 + s*f - i

# vac_seeking_probability = (alpha_i - d.loc[n, 'Distance'] * beta_i).clip(min = gamma_i)

# delay_cdf = (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**(- s * f + i) - (1 - p_vac.loc[p_vac['Vaccine'] == a, 'Vaccination Delay Geom p'])**(- (s-1)*f + i)

(a) Di,t is the vaccine demand volume at location i on day t <br>
(b) N is the set of birth-nodes assigned to the health facility location i <br>
(c) A is the set of antigens (eg. BCG, Pentavalent, Rotavirus etc) as per National
Immunization Schedule.
(d) s is variable for session relative to time t when child became eligible for vaccination,
s = 0 is the current session, s = −1 is previous session and so on.
(e) f is the frequency of vaccination. If once every 28 days, then f = 28.
(f) Xn,t is the number of births at birth node n on day t. Xn,t ∼ P oi( λn
365 ) where λn is
annual number of births at node n
(g) vac starta is the start date since date of birth when the child is eligible for antigen a
(h) α + βd is Vaccination Seeking probability ∈ [0.4, 0.8], d being distance from nearest
Health Facility
(i) 1/1+eθ is the Caregiver’s Location Choice probability of vaccinating at AWC, θ being
AWC/PHC distance ratio.
(j) 1 − (1 − pvac)k represents cdf P [delay ≤ k] of Geometric model of Caregiver’s
Vaccination Delay, pvac being the vaccine-level delay parameter
(k) δb is the fraction of caregivers who delay vaccination due to lack of continuous vaccine
availability
(l) I is an indicator variable such that I = 1 if s = 0 and I = 0 otherwise. I allows
inclusion of δb proportion of caregivers who vaccinate their children immediately
without delay
(m) vac vola is the Wastage adjusted average packed volume secondary package for antigen a
