==> scripts/09 MLE.ipynb <==
import pandas as pd

import statsmodels.api as sm

import numpy as np

df = pd.read_csv(r"C:\Users\Rohan Bapat\Documents\Projects\Immunization Supply Chain\immunization_supply_chain\data\choices_made_df5.csv")

df.columns

df['y'] = df[['ChosenSS', 'ChosenPHC', 'ChosenNO']].apply(lambda x: np.argmax(x), axis = 1)

x_cols = ['dist_phc', 'dist_ss', 'sociodemographic', 'prior_unserved_phc', 'prior_unserved_ss', 'log_days_since_due_date']

mlogit_mod = sm.MNLogit(df['y'], sm.add_constant(df[x_cols]))
mlogit_res = mlogit_mod.fit()
print(mlogit_res.params)

anes_data = sm.datasets.anes96.load_pandas()
anes_exog = anes_data.exog
anes_exog = sm.add_constant(anes_exog)

anes_data.data

print(anes_data.exog.head())
print(anes_data.endog.head())
