==> scripts/09a Create choice dataset.ipynb <==
import SimulateDemand
import pandas as pd
import numpy as np

scen_params_setting = {'scen1': {'ss_mode': 'periodic', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':0},
                       'scen2': {'ss_mode': 'periodic', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0.2 , 'ccedisruption':0},
                       'scen3': {'ss_mode': 'periodic', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':4/365},
                       'scen4': {'ss_mode': 'periodic-2device', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':0},
                       'scen5': {'ss_mode': 'periodic-2device', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0.2 , 'ccedisruption':0},
                       'scen6': {'ss_mode': 'periodic-2device', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':4/365},
                       'scen7': {'ss_mode': 'periodic-2device', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0.2 , 'ccedisruption':4/365},
                       'scen8': {'ss_mode': 'continuous', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':0},
                       'scen9': {'ss_mode': 'continuous', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0.2 , 'ccedisruption':0},
                       'scen10': {'ss_mode': 'continuous', 'ss_replenishment_source': 'phc', 'replenishmentdisruption': 0 , 'ccedisruption':4/365},
                       'scen11': {'ss_mode': 'continuous', 'ss_replenishment_source': 'dvs', 'replenishmentdisruption': 0 , 'ccedisruption':0},
                       'scen12': {'ss_mode': 'continuous', 'ss_replenishment_source': 'dvs', 'replenishmentdisruption': 0.2 , 'ccedisruption':0},
                       'scen13': {'ss_mode': 'continuous', 'ss_replenishment_source': 'dvs', 'replenishmentdisruption': 0 , 'ccedisruption':4/365},
                       'scen14': {'ss_mode': 'continuous', 'ss_replenishment_source': 'dvs', 'replenishmentdisruption': 0.2 , 'ccedisruption':4/365}
                      }

i=0

choices_df = pd.DataFrame()

for scen, scen_params in scen_params_setting.items():
    scen_df = SimulateDemand.main(i, scen, scen_params['ss_mode'], scen_params['ss_replenishment_source'], scen_params['replenishmentdisruption'], scen_params['ccedisruption'])
    choices_df = choices_df.append(scen_df)
    print(scen, scen_df.shape[0])
    i+=1

choices_df.to_csv(r"C:\Users\Rohan Bapat\Documents\Projects\Immunization Supply Chain\immunization_supply_chain\data\09 choices for MLE smallerrors.csv", index = False)

choices_df[choices_df['ChosenNO']==1].shape[0]

choices_sample = choices_df[choices_df['ChosenNO']==1].sample(300000)

choices_sample = choices_sample.append(choices_df[choices_df['ChosenNO']!=1].sample(10000, replace = True))

choices_sample.shape[0]

choices_sample.to_csv(r"C:\Users\Rohan Bapat\Documents\Projects\Immunization Supply Chain\immunization_supply_chain\data\09 choices for MLE sample v2.csv", index = False)

choices_sample.columns
